Added Arduino and AVR MCU support.

+++ emJSON/json.h
@@ -63,7 +63,7 @@ int json_set_float(json_t *obj, char *key, float value);
 
 // String-related functions
 int json_strcpy(char *dest, json_t *obj);
-int json_strlen(json_t *obj);
+size_t json_strlen(json_t *obj);
 
 // Buffer and memory management functions
 int json_replace_buffer(json_t *obj, void *new_buf, size_t size);
@@ -71,9 +71,9 @@ int json_double_table(json_t *obj);
 json_t json_copy(void *dest_buf, json_t *obj);
 
 // Other utility functions
-int json_table_size(json_t *obj);
-int json_count(json_t *obj);
-int json_buffer_size(json_t *obj);
+size_t json_table_size(json_t *obj);
+size_t json_count(json_t *obj);
+size_t json_buffer_size(json_t *obj);
 
 #ifdef __cplusplus
 }
